requirements.txt
requests>=2.31.0,<3.0.0
urllib3>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
flask>=3.0.0,<4.0.0
flask-cors>=4.0.0,<5.0.0
gunicorn>=20.1.0,<22.0.0
